refact: organiza os scrapers em diretórios
O projeto vai ter mais de 1 scraper, então vamos organizá-los em
diretórios. E poderemos ter mais de 1 scraper por domínio pois podemos
querer extrair dados diferentes de para propósitos diferentes. Ex: num
momento coletar matérias disponíveis sobre um time e depois extrair
conteúdo de cada matéria separadamente em paralelo, por exemplo.
